backend/requirements.txt
fastapi
uvicorn
pymongo
pydantic
python-dotenv
python-multipart
fpdf
boto3
PyJWT
